+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- command line options --help and --version
+
 ## [0.0.33] - 2019-06-18
 
 ### Added